cast string to integer, doh!

+++ docker/as/client/required/scripts/rtf_client.groovy
@@ -32,8 +32,8 @@ def time = new Date();
 URI sourceUri;
 def scriptDir = new File(sourceUri).parent;
 
-def batchBegin = env["BATCH_BEGIN"] ?: 0;
-def batchEnd = env["BATCH_END"] ?: 9999;
+def batchBegin = (env["BATCH_BEGIN"] ?: 0) as Integer;
+def batchEnd = (env["BATCH_END"]?: 9999) as Integer;
 def batchOffset = env["BATCH_OFFSET"] ?:  new Random().nextInt(batchEnd - batchBegin);
 
 def rtfDataSource = new BasicDataSource();
